Fix some warnings

+++ concurrent_map.go
@@ -33,6 +33,7 @@ func New[K comparable, V any]() *ConcurrentMap[K, V] {
 	for i := 0; i < SHARD_COUNT; i++ {
 		shared[i] = &ConcurrentMapShared[K, V]{items: make(map[K]V)}
 	}
+
 	return &ConcurrentMap[K, V]{
 		shards:   shared,
 		sharding: genHasher[K](),
@@ -45,13 +46,14 @@ func NewWithCustomShardingFunction[K comparable, V any](sharding func(key K) uin
 	for i := 0; i < SHARD_COUNT; i++ {
 		shared[i] = &ConcurrentMapShared[K, V]{items: make(map[K]V)}
 	}
+
 	return ConcurrentMap[K, V]{
 		shards:   shared,
 		sharding: sharding,
 	}
 }
 
-// GetShard returns shard under given key
+// GetShard returns shard under given key.
 func (m ConcurrentMap[K, V]) GetShard(key K) *ConcurrentMapShared[K, V] {
 	return m.shards[uint(m.sharding(key))%uint(SHARD_COUNT)]
 }
@@ -77,10 +79,10 @@ func (m ConcurrentMap[K, V]) Set(key K, value V) {
 // Callback to return new element to be inserted into the map
 // It is called while lock is held, therefore it MUST NOT
 // try to access other keys in same map, as it can lead to deadlock since
-// Go sync.RWLock is not reentrant
+// Go sync.RWLock is not reentrant.
 type UpsertCb[K comparable, V any] func(exist bool, valueInMap V, newValue V) V
 
-// Insert or Update - updates existing element or inserts a new one using UpsertCb
+// Insert or Update - updates existing element or inserts a new one using UpsertCb.
 func (m ConcurrentMap[K, V]) Upsert(key K, value V, cb UpsertCb[K, V]) (res V) {
 	shard := m.GetShard(key)
 	shard.Lock()
@@ -88,6 +90,7 @@ func (m ConcurrentMap[K, V]) Upsert(key K, value V, cb UpsertCb[K, V]) (res V) {
 	res = cb(ok, v, value)
 	shard.items[key] = res
 	shard.Unlock()
+
 	return res
 }
 
@@ -101,6 +104,7 @@ func (m ConcurrentMap[K, V]) SetIfAbsent(key K, value V) bool {
 		shard.items[key] = value
 	}
 	shard.Unlock()
+
 	return !ok
 }
 
@@ -112,6 +116,7 @@ func (m ConcurrentMap[K, V]) Get(key K) (V, bool) {
 	// Get item from shard.
 	val, ok := shard.items[key]
 	shard.RUnlock()
+
 	return val, ok
 }
 
@@ -124,10 +129,11 @@ func (m ConcurrentMap[K, V]) Count() int {
 		count += len(shard.items)
 		shard.RUnlock()
 	}
+
 	return count
 }
 
-// Looks up an item under specified key
+// Looks up an item under specified key.
 func (m ConcurrentMap[K, V]) Has(key K) bool {
 	// Get shard
 	shard := m.GetShard(key)
@@ -135,6 +141,7 @@ func (m ConcurrentMap[K, V]) Has(key K) bool {
 	// See if element is within shard.
 	_, ok := shard.items[key]
 	shard.RUnlock()
+
 	return ok
 }
 
@@ -148,12 +155,12 @@ func (m ConcurrentMap[K, V]) Remove(key K) {
 }
 
 // RemoveCb is a callback executed in a map.RemoveCb() call, while Lock is held
-// If returns true, the element will be removed from the map
+// If returns true, the element will be removed from the map.
 type RemoveCb[K comparable, V any] func(key K, v V, exists bool) bool
 
 // RemoveCb locks the shard containing the key, retrieves its current value and calls the callback with those params
 // If callback returns true and element exists, it will remove it from the map
-// Returns the value returned by the callback (even if element was not present in the map)
+// Returns the value returned by the callback (even if element was not present in the map).
 func (m ConcurrentMap[K, V]) RemoveCb(key K, cb RemoveCb[K, V]) bool {
 	// Try to get shard.
 	shard := m.GetShard(key)
@@ -164,10 +171,11 @@ func (m ConcurrentMap[K, V]) RemoveCb(key K, cb RemoveCb[K, V]) bool {
 		delete(shard.items, key)
 	}
 	shard.Unlock()
+
 	return remove
 }
 
-// Pop removes an element from the map and returns it
+// Pop removes an element from the map and returns it.
 func (m ConcurrentMap[K, V]) Pop(key K) (v V, exists bool) {
 	// Try to get shard.
 	shard := m.GetShard(key)
@@ -175,6 +183,7 @@ func (m ConcurrentMap[K, V]) Pop(key K) (v V, exists bool) {
 	v, exists = shard.items[key]
 	delete(shard.items, key)
 	shard.Unlock()
+
 	return v, exists
 }
 
@@ -183,7 +192,7 @@ func (m ConcurrentMap[K, V]) IsEmpty() bool {
 	return m.Count() == 0
 }
 
-// Used by the Iter & IterBuffered functions to wrap two variables together over a channel,
+// Used by the Iter & IterBuffered functions to wrap two variables together over a channel,.
 type Tuple[K comparable, V any] struct {
 	Key K
 	Val V
@@ -198,6 +207,7 @@ func (m ConcurrentMap[K, V]) IterBuffered() <-chan Tuple[K, V] {
 	}
 	ch := make(chan Tuple[K, V], total)
 	go fanIn(chans, ch)
+
 	return ch
 }
 
@@ -213,7 +223,7 @@ func (m ConcurrentMap[K, V]) Clear() {
 // It returns once the size of each buffered channel is determined,
 // before all the channels are populated using goroutines.
 func snapshot[K comparable, V any](m ConcurrentMap[K, V]) (chans []chan Tuple[K, V]) {
-	//When you access map items before initializing.
+	// When you access map items before initializing.
 	if len(m.shards) == 0 {
 		panic(`cmap.ConcurrentMap is not initialized. Should run New() before usage.`)
 	}
@@ -235,10 +245,11 @@ func snapshot[K comparable, V any](m ConcurrentMap[K, V]) (chans []chan Tuple[K,
 		}(index, shard)
 	}
 	wg.Wait()
+
 	return chans
 }
 
-// fanIn reads elements from channels `chans` into channel `out`
+// fanIn reads elements from channels `chans` into channel `out`.
 func fanIn[K comparable, V any](chans []chan Tuple[K, V], out chan Tuple[K, V]) {
 	wg := sync.WaitGroup{}
 	wg.Add(len(chans))
@@ -254,7 +265,7 @@ func fanIn[K comparable, V any](chans []chan Tuple[K, V], out chan Tuple[K, V]) 
 	close(out)
 }
 
-// Items returns all items as map[string]V
+// Items returns all items as map[string]V.
 func (m ConcurrentMap[K, V]) Items() map[K]V {
 	tmp := make(map[K]V)
 
@@ -269,7 +280,7 @@ func (m ConcurrentMap[K, V]) Items() map[K]V {
 // Iterator callbacalled for every key,value found in
 // maps. RLock is held for all calls for a given shard
 // therefore callback sess consistent view of a shard,
-// but not across the shards
+// but not across the shards.
 type IterCb[K comparable, V any] func(key K, v V)
 
 // Callback based iterator, cheapest way to read
@@ -285,7 +296,7 @@ func (m ConcurrentMap[K, V]) IterCb(fn IterCb[K, V]) {
 	}
 }
 
-// Keys returns all keys as []string
+// Keys returns all keys as []string.
 func (m ConcurrentMap[K, V]) Keys() []K {
 	count := m.Count()
 	ch := make(chan K, count)
@@ -313,6 +324,7 @@ func (m ConcurrentMap[K, V]) Keys() []K {
 	for k := range ch {
 		keys = append(keys, k)
 	}
+
 	return keys
 }
 
@@ -325,6 +337,7 @@ func (m ConcurrentMap[K, V]) MarshalJSON() ([]byte, error) {
 	for item := range m.IterBuffered() {
 		tmp[item.Key] = item.Val
 	}
+
 	return json.Marshal(tmp)
 }
 
@@ -341,17 +354,19 @@ func (m *ConcurrentMap[K, V]) UnmarshalJSON(b []byte) (err error) {
 	for key, val := range tmp {
 		m.Set(key, val)
 	}
+
 	return nil
 }
 
 const (
-	// hash input allowed sizes
+	// hash input allowed sizes.
 	byteSize = 1 << iota
 	wordSize
 	dwordSize
 	qwordSize
 	owordSize
 )
+
 const (
 	prime1 uint64 = 11400714785074694791
 	prime2 uint64 = 14029467366897019727
@@ -369,12 +384,15 @@ func round(acc, input uint64) uint64 {
 	acc += input * prime2
 	acc = rol31(acc)
 	acc *= prime1
+
 	return acc
 }
+
 func mergeRound(acc, val uint64) uint64 {
 	val = round(0, val)
 	acc ^= val
 	acc = acc*prime1 + prime4
+
 	return acc
 }
 func rol1(x uint64) uint64  { return bits.RotateLeft64(x, 1) }
@@ -386,9 +404,9 @@ func rol23(x uint64) uint64 { return bits.RotateLeft64(x, 23) }
 func rol27(x uint64) uint64 { return bits.RotateLeft64(x, 27) }
 func rol31(x uint64) uint64 { return bits.RotateLeft64(x, 31) }
 
-// xxHash implementation for known key type sizes, minimal with no branching
+// xxHash implementation for known key type sizes, minimal with no branching.
 var (
-	// byte hasher, key size -> 1 byte
+	// byte hasher, key size -> 1 byte.
 	byteHasher = func(key uint8) uintptr {
 		h := prime5 + 1
 		h ^= uint64(key) * prime5
@@ -398,9 +416,10 @@ var (
 		h ^= h >> 29
 		h *= prime3
 		h ^= h >> 32
+
 		return uintptr(h)
 	}
-	// word hasher, key size -> 2 bytes
+	// word hasher, key size -> 2 bytes.
 	wordHasher = func(key uint16) uintptr {
 		h := prime5 + 2
 		h ^= (uint64(key) & 0xff) * prime5
@@ -412,9 +431,10 @@ var (
 		h ^= h >> 29
 		h *= prime3
 		h ^= h >> 32
+
 		return uintptr(h)
 	}
-	// dword hasher, key size -> 4 bytes
+	// dword hasher, key size -> 4 bytes.
 	dwordHasher = func(key uint32) uintptr {
 		h := prime5 + 4
 		h ^= uint64(key) * prime1
@@ -424,9 +444,10 @@ var (
 		h ^= h >> 29
 		h *= prime3
 		h ^= h >> 32
+
 		return uintptr(h)
 	}
-	// qword hasher, key size -> 8 bytes
+	// qword hasher, key size -> 8 bytes.
 	qwordHasher = func(key uint64) uintptr {
 		k1 := key * prime2
 		k1 = bits.RotateLeft64(k1, 31)
@@ -438,6 +459,7 @@ var (
 		h ^= h >> 29
 		h *= prime3
 		h ^= h >> 32
+
 		return uintptr(h)
 	}
 	float32Hasher = func(key float32) uintptr {
@@ -450,6 +472,7 @@ var (
 		h ^= h >> 29
 		h *= prime3
 		h ^= h >> 32
+
 		return uintptr(h)
 	}
 	float64Hasher = func(key float64) uintptr {
@@ -462,6 +485,7 @@ var (
 		h ^= h >> 29
 		h *= prime3
 		h ^= h >> 32
+
 		return uintptr(h)
 	}
 )
@@ -551,13 +575,14 @@ func genHasher[K comparable]() func(K) uintptr {
 		hasher = func(key K) uintptr {
 			k := *(*uint64)(unsafe.Pointer(&key))
 			h := prime5 + 4
-			h ^= uint64(k) * prime1
+			h ^= k * prime1
 			h = bits.RotateLeft64(h, 23)*prime2 + prime3
 			h ^= h >> 33
 			h *= prime2
 			h ^= h >> 29
 			h *= prime3
 			h ^= h >> 32
+
 			return uintptr(h)
 		}
 	case complex128:
@@ -616,5 +641,6 @@ func genHasher[K comparable]() func(K) uintptr {
 			}
 		}
 	}
+
 	return hasher
 }
